fastapi-ssh-service/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
python-memcached
asyncssh
kubernetes
asyncio
python-dotenv